Initial commit of the manual_classifier
This is used to produce data sets for the greeter project by
displaying a set of user-selected images to the user and allowing
them to manually specify who of Jacob, Emma etc are present in the
image.  The images themselves are converted to cv2 format and are
alsoshrunkdown in size a bit before being stored in a h5py file.
This is in no way ready for use though - the database part isn't
ready yet and the images (both as rendered in the UI and as
rendered in the compressed format) are not scaled properly.